Move entryPoint to webpack
