+++ packages/webpack-common/src/webpack.js
@@ -7,7 +7,6 @@ import path from 'path'
 export {
   addPlugins,
   customConfig,
-  entryPoint,
   performance,
   resolveAliases,
   setContext,
@@ -27,40 +26,6 @@ function addPlugins (plugins) {
 
 function customConfig (wpConfig) {
   return () => wpConfig
-}
-
-/**
- * Adds one or multiple entry points. If the parameter is not an object the
- * entry point(s) will be added to the default chunk named `main`.
- *
- * @param {object|string[]|string} entry
- * @see https://webpack.github.io/docs/configuration.html#entry
- */
-function entryPoint (entry) {
-  return () => ({
-    entry: normalizeEntry(entry)
-  })
-}
-
-function normalizeEntry (entry) {
-  if (Array.isArray(entry)) {
-    return {
-      main: entry
-    }
-  } else if (typeof entry === 'string') {
-    return {
-      main: [ entry ]
-    }
-  } else if (typeof entry === 'object') {
-    Object.keys(entry).forEach((entryName) => {
-      if (!Array.isArray(entry[entryName])) {
-        entry[entryName] = [ entry[entryName] ]
-      }
-    })
-    return entry
-  } else {
-    throw new Error(`Expected entry point to be object, array or string. Instead got: ${entry}`)
-  }
 }
 
 /**
